Converting Point float to point<T> Some fixes on sprite setters/getters

+++ src/Engine/Engine.h
@@ -101,7 +101,7 @@ public:
 private:
 	void Update();
 	void Draw();
-	float lastTick = 0;
+	Uint32 lastTick = 0;
 
 	float DeltaTime()
 	{
